ai-trends-app/requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-Session==0.5.0
flask-cors==4.0.0

# OpenAI and LangChain - FULLY UPDATED FOR PYTHON 3.13
openai>=1.10.0
tiktoken>=0.7.0
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-community>=0.0.38

# Web Scraping
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0

# Data Processing
PyYAML==6.0.1
python-dotenv==1.0.1
pandas>=2.2.0
numpy>=1.26.4

# Tavily Search API
tavily-python==0.3.3

# Utilities
Werkzeug==3.0.1
openpyxl==3.1.2

# Password Hashing
flask-bcrypt==1.0.1

# CSRF Protection
flask-wtf==1.2.1

# Production Web Server
gunicorn==21.2.0
